% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_betas.R
\name{allpos_helper}
\alias{allpos_helper}
\title{All possible regression internal}
\usage{
allpos_helper(model)
}
\arguments{
\item{model}{An object of class \code{lm}.}
}
\description{
Internal function for all possible regression.
}
