% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdata.R
\name{checkdata}
\alias{checkdata}
\title{Data check for errors}
\usage{
checkdata(x)
}
\arguments{
\item{x}{is a data frame}
}
\description{
Preliminary check of data frame for missing values, wrong format, outliers.
}
\examples{
data(macroKZ)
checkdata(macroKZ)
}
