% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_betas.R
\name{coeff_length}
\alias{coeff_length}
\title{Coefficients length}
\usage{
coeff_length(predicts, gap)
}
\arguments{
\item{predicts}{Name of the predictors in the model.}
\item{gap}{A numeric vector.}
}
\value{
no return value
}
\description{
Returns the length of the coefficient names.
}
