% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsel.R
\name{corsel}
\alias{corsel}
\title{Multicollinearity test}
\usage{
corsel(x, thrs, num)
}
\arguments{
\item{x}{is a numeric vector or matrix}

\item{thrs}{threshold set to calculate correlation above}

\item{num}{logical}
}
\value{
A matrix of numerical and logical values of correlation between variables.
}
\description{
multicollinearity is the occurence of high interrelations among two or more independent variables in a multiple regression.
}
