% Generated by roxygen2: do not edit by hand
\name{HP}
\alias{HP}
\title{Hodrick-Prescott filter}
\description{
Hodrick-Prescott filter is a data smoothing technique that removes trending in time series data frame
}
\usage{
HP(x, freq = NULL, type = c("lambda", "frequency"), drift = FALSE)
}
\arguments{
  \item{x}{time-series vector}
  \item{freq}{integer}
  \item{type}{character, indicating the filter type}
  \item{drift}{logical}
}
\examples{
data (macroKZ)
HP(macroKZ[,2])
}

