% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroKZ.R
\docType{data}
\name{macroKZ}
\alias{macroKZ}
\title{macroKZ dataset}
\source{
Bureau of National statistics, Agency for Strategic planning and reforms of the Republic of Kazakhstan
}
\usage{
macroKZ
}
\description{
A time series data frame of 48 quarterly observations of 50 macroeconomic and 10 financial parameters for 2010-2021 period.
}
\references{
The Agency of the Republic of Kazakhstan for Regulation and Development of Financial Market
}
\keyword{datasets}
