% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fout_ggtheme.R
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 theme}
\usage{
my_theme(base_size = 12, base_family = "sans", ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
A custom ggplot2 theme!
}
\examples{
cat_function()
}
\keyword{cats}
