% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_size.R
\name{opt_size}
\alias{opt_size}
\title{Necessary size of the time-series dataset}
\usage{
opt_size(model)
}
\arguments{
\item{model}{is a linear regression model a class \code{lm}.}
}
\value{
A numerical value of the number of models.
}
\description{
Estimates number of models generated from given number of regressors {x}
}
