% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{pct4}
\alias{pct4}
\title{Transforming time-series data to stationary}
\usage{
pct4(x)
}
\arguments{
\item{x}{time-series vector(s)}
}
\value{
A vector of length shorter than the original vector. Returned value is a numeric.
}
\description{
Percent change is a change between a term and its lagged value for prior period, %
}
