% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_reg.R
\name{vif_reg}
\alias{vif_reg}
\title{VIF by variable}
\usage{
vif_reg(model)
}
\arguments{
\item{model}{is a linear regression model}
}
\value{
A data frame with columns for variable and VIF result. Textual result for VIF test presents whether variables pass the test or not.
}
\description{
Calculates the variation inflation factors of all predictors in regression models
}
\references{
Petrie, Adam. Published 2020-02-21. regclass package
}
