% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf.R
\name{adf}
\alias{adf}
\title{Augmented Dickey Fuller Test}
\usage{
adf(x, k = trunc((length(x) - 1)^(1/3)))
}
\arguments{
\item{x}{time-series vector}

\item{k}{the lag order to calculate the test statistic.}
}
\description{
ADF test are used to test stationarity of a time-series data
}
\examples{
data(macroKZ)
adf(macroKZ)
}
\references{
Trapletti, A., Augmented Dickey-Fuller Test
Trapletti, A., KPSS Test for Stationarity
}
