% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{HP}
\alias{HP}
\title{Hodrick-Prescott filter}
\usage{
HP(x, freq = NULL, type = c("lambda", "frequency"), drift = FALSE)
}
\arguments{
\item{x}{time-series vector}

\item{freq}{integer}

\item{type}{character, indicating the filter type}

\item{drift}{logical}
}
\value{
A "mFilter" object of the mFilter package.
}
\description{
Hodrick-Prescott filter is a data smoothing technique that removes trending in time series data frame
}
