% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macroKZ.R
\docType{data}
\name{macroKZ}
\alias{macroKZ}
\title{macroKZ dataset}
\format{
A time series data frame of 50 quarterly observations of 49 macroeconomic and 11 financial parameters for 2010-2022 period.
}
\source{
Bureau of National statistics, Agency for Strategic planning and reforms of the Republic of Kazakhstan
}
\usage{
macroKZ
}
\description{
macroKZ dataset
}
\references{
The Agency of the Republic of Kazakhstan for Regulation and Development of Financial Market
}
\keyword{datasets}
